First Prototype: Logo and Banner image
Header is the logo and banner image, as in the first prototype in the email.
The navigation bar however extends accross the whole pages. I think having
it only extend partially across as in the prototype, and have content on
the left hand side blurs the separation between the header and the content.